2nd and 3rd Practice with EspNOW
In this 2nd and 3rd practice, we needed to create a connection with EspNOW, to connect two Esp32.
Specifically, in the 2nd practice, we needed to turn on and off the backlight of the i2c display. 
Meanwhile in the 3rd practice, we needed to transmit a message, which contains my name and transmit it, to the other Esp32. When the name appears, also a counter will appear, which after 10 seconds, the sent name changes again to my name (or the partner name if needed).
In the “transmitter message (20x4)” and in the “receiver message (20x4)”, I added more things. In addition to turning on and off the backlight of the i2c screen and sending my name, I made it possible to modify the message to be sent, with the facility to move from left to right, and viceversa, in order to modify the message with a maximum of 20 letters.